* Compile fix

+++ libcluon/include/cluon/Envelope.hpp
@@ -54,7 +54,7 @@ inline std::pair<bool, cluon::data::Envelope> extractEnvelope(std::istream &in) 
                 && (0xA4 == static_cast<uint8_t>(buffer[1]))) {
                 const uint32_t LENGTH{le32toh(*reinterpret_cast<uint32_t*>(&buffer[1])) >> 8};
                 buffer.reserve(LENGTH);
-                retVal = LENGTH == in.readsome(&buffer[0], static_cast<std::streamsize>(LENGTH));
+                retVal = static_cast<int32_t>(LENGTH) == in.readsome(&buffer[0], static_cast<std::streamsize>(LENGTH));
                 if (retVal) {
                     std::stringstream sstr;
                     sstr.rdbuf()->pubsetbuf(&buffer[0], static_cast<std::streamsize>(LENGTH)); // Avoid duplicating the read data.
